+++ blockstack-sdk/src/main/java/org/blockstack/android/sdk/ui/ConnectHelpFragment.kt
@@ -1,0 +1,30 @@
+package org.blockstack.android.sdk.ui
+
+import android.app.Activity.RESULT_OK
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.Button
+import androidx.fragment.app.Fragment
+import org.blockstack.android.sdk.R
+
+
+class ConnectHelpFragment : Fragment() {
+
+    override fun onCreateView(
+            inflater: LayoutInflater, container: ViewGroup?,
+            savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_connect_help, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        view.findViewById<Button>(R.id.button_get_started).setOnClickListener {
+            requireActivity().setResult(RESULT_OK)
+            requireActivity().finish()
+        }
+    }
+}