+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unrelease
 ### Added
 - additional parameters `sendToSignIn`, `appDetails` for `BlockstackSignIn`
+- `ConnectBottemSheetDialogFragment` and `SignInProvider` to show user on-boarding
 ### Changed
 - use app.blockstack.org (connect) instead of browser.blockstack.org for sign in.